1. Preliminary attempt at https://blueprints.launchpad.net/yade/+spec/c++-numpy-arrays (for postprocessing and similar), example in eudoxos.testNumpy 2. SpherePack::makeCloud can be asked for negative number of spheres (= as much as possible)

+++ pkg/dem/DataClass/SpherePack.hpp
@@ -50,8 +50,8 @@ public:
 	void toFile(const string file) const;
 	void fromSimulation();
 
-	// random generation
-	long makeCloud(Vector3r min, Vector3r max, Real rMean, Real rFuzz, size_t num, bool periodic=false);
+	// random generation; if num<0, insert as many spheres as possible
+	long makeCloud(Vector3r min, Vector3r max, Real rMean, Real rFuzz, int num, bool periodic=false);
 
 	// periodic repetition
 	void cellRepeat(Vector3<int> count);
